fix: utils state import into content script
